Crate icu_provider_cldr

icu_provider_cldr contains implementations of the ICU4X data provider interface based on the JSON files shipped by CLDR. Create a CldrPaths and then pass it into CldrJsonDataProvider.

This crate contains two implementations of CldrPaths:

CldrPathsLocal points to local copies of the CLDR JSON repositories.

CldrPathsDownload downloads and caches the CLDR JSON repositories. Requires the “download” feature.

Important: This data provider implementation is not optimized for production use. It is much more efficient if you use FsDataProvider instead.

Re-exports

pub use transform::CldrJsonDataProvider;

Modules

download

transform

This module contains the core transformer code from CLDR JSON to ICU4X Data Provider.

Structs

CldrPathsAllInOne

An implementation of CldrPaths for one combined local CLDR JSON directory.

CldrPathsLocal

An implementation of CldrPaths for multiple separate local CLDR JSON directories per component.

Enums

CldrError

Traits

CldrPaths

Trait returning filesystem paths to CLDR JSON resource directories. The fields should be Ok if present. They default to Err when not present.

KeyedDataProvider

A data provider whose supported keys are known statically at compile time.