Module icu_provider_cldr::transform

This module contains the core transformer code from CLDR JSON to ICU4X Data Provider.

Every ICU4X component should have its own private submodule and then export the types from here.

Structs

AliasesProvider

A data provider reading from CLDR JSON likely subtags rule files.

CldrJsonDataProvider

DatePatternsProvider

A data provider reading from CLDR JSON dates files.

DateSkeletonPatternsProvider

A data provider reading from CLDR JSON dates files.

DateSymbolsProvider

A data provider reading from CLDR JSON dates files.

LikelySubtagsProvider

A data provider reading from CLDR JSON likely subtags rule files.

ListProvider

A data provider reading from CLDR JSON list rule files.

NumbersProvider

A data provider reading from CLDR JSON plural rule files.

PluralsProvider

A data provider reading from CLDR JSON plural rule files.

Functions

get_all_cldr_keys

Returns a list of all ResourceKeys that this provider can produce.