Crate icu_provider_cldr

icu_provider_cldr is one of the ICU4X components.

It contains implementations of the DataProvider interface based on the JSON files shipped by CLDR. You create a CldrPaths and then pass it into CldrJsonDataProvider.

This crate contains two implementations of CldrPaths:

CldrPathsLocal points to local copies of the CLDR JSON repositories.

CldrPathsDownload downloads and caches the CLDR JSON repositories. Requires the "download" feature.

Important: This data provider implementation is not optimized for production use. It is much more efficient if you use FsDataProvider instead.

Re-exports

pub use transform::CldrJsonDataProvider;

Modules

download

transform

Structs

CldrPathsLocal
Implementation of CldrPaths for data directories already downloaded.

Enums

CldrError

Traits

CldrPaths
Trait returning filesystem paths to CLDR JSON resource directories. The fields should be Ok if present. They default to Err when not present.